backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
firebase-admin==6.2.0
redis==5.0.1
stripe==7.8.0
razorpay==1.3.0
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.9
httpx==0.25.2
websockets==12.0
python-socketio==5.10.0
cors==1.0.1
gunicorn==21.2.0